[RIKU-8] refactor: NavBar shared 폴더로 분리 및 메인 layout 라우터 재설정
