+++ src/app/App.tsx
@@ -4,7 +4,8 @@ import MyPage from '../components/Main/MyPage';
 import FlashRunMake from '../components/FlashRun/FlashRunMake';
 import FlashRunDetail from '../components/FlashRun/FlashRunDetail';
 import LoginPage from '../pages/LoginPage';
-import TabNavigationUI from '../components/TabNavigationUI';
+import MainTabLayout from '@app/layouts/MainTabLayout';
+import RankingPage from '@pages/RankingPage';
 
 import NewMain from '../components/MainPage/NewMain';
 import AdminPage from '../components/AdminPage/AdminPage';
@@ -67,8 +68,14 @@ function App() {
             <Route path="/make/event" element={<EventMake />} />
             <Route path="/make/training" element={<TrainingMake />} />
 
-            <Route path="/tab/*" element={<TabNavigationUI />} />
-            <Route path="/main" element={<NewMain />} />
+            {/* 메인화면, 일정, 순위, 마이페이지 페이지를 이용해서 오고가는 MainTabLayout 페이지 */}
+            <Route path="/tab" element={<MainTabLayout />}>
+              <Route index element={<NewMain />} /> {/* /tab → 기본 탭 */}
+              <Route path="main" element={<NewMain />} />
+              <Route path="schedule-page" element={<SchedulePage />} />
+              <Route path="ranking-page" element={<RankingPage />} />
+              <Route path="my-page" element={<MyPage />} />
+            </Route>
             <Route path="/:runType" element={<RunList />} />
 
             <Route path="/regular/edit/:postId" element={<NewRegularRunEdit />} />
