+++ src/components/ActionBar.tsx
@@ -11,7 +11,7 @@ const ActionBar: React.FC = () => {
 
   return (
     <div className="fixed top-0 inset-x-0 mx-auto flex justify-between items-center max-w-[430px] w-full h-[56px] border-t-[1.5px] z-[1000] border-kuDarkGreen bg-kuDarkGreen">
-      {/* 상단 액션 바(TabNavigationUI에서 하단의 NavBar와 함께 계속 떠있을 것임) */}
+      {/* 상단 액션 바(MainTabLayout에서 하단의 NavBar와 함께 계속 떠있을 것임) */}
       {/* 아이콘 2개를 양쪽에 정렬한다 */}
       <div onClick={handleMain} className="cursor-pointer">
         <object data={rikuLogo_text} className="ml-8 w-auto h-auto pointer-events-none" />{' '}
