+++ src/components/TabNavigationUI.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import NavBar from './NavBar';
+import NavBar from '@shared/ui/NavBar';
 import MyPage from './Main/MyPage';
 import SchedulePage from './Main/SchedulePage';
-import RankingPage from './Main/RankingPage';
+import RankingPage from '../pages/RankingPage';
 import NewMain from './MainPage/NewMain';
 import ActionBar from '../components/ActionBar';
