+++ src/features/ranking/components/EventSection.tsx
