+++ src/features/ranking/components/EventSectionTopBanner.tsx
