+++ src/features/ranking/components/RankingSection.tsx
