+++ src/typings.d.ts
@@ -18,11 +18,6 @@ declare module '*.svg' {
   export default value;
 }
 
-declare module '*.svg' {
-  const value: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
-  export default value;
-}
-
 declare module '*.mp4' {
   const src: string;
   export default src;
